LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_sequence_at_implicit_release_on_click_pen-expected.txt

Pointer Event: Event sequence at implicit release on click

When a captured pointer is implicitly released after a click, the boundary events should follow the lostpointercapture event.

Click or tap on Black.
Click or tap on Green.

PASS pen Boundary events are emitted after lostpointercapture
FAIL No extra events are emitted assert_equals: expected "" but got "pointerover, pointerenter, pointerout, pointerleave"
